test(baseBallModelTest) : 베이스볼 모델 관련 테스트로 폴더이동
